Updated JoJo WotD list and expanded the json
JoJoWotD files now contain all WotD from 2013/10/12 to 2016/12/09
The jojowotd.txt file now contains only the WotD, without the
timestampts.
The jojowotd.json file now includes the WotD with spaces as they appear
in the HTML file, all datereleasedUTC fields are filled with their
correct dates and the preGit field, well is if the WotD was released
before the migration to Git :)
The old jojowotd.txt file is now known as jojowotd_pregit.txt, it is
needed for the pre-git dates.
Full new format:
{ "wotds": [
{ "wotd" : "theword",
"definition" : "the definition",
"dateReleasedUTC": "Date and Time in UTC format",
"length": #,
"properWotD": "The word with spaces",
"preGit": boolean }
]}